+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This section contains changes that have been committed but not yet released.
 
 - Added Outbox support
 - Added support for Component CRUD
+- Added support for Neural Categorizer
 
 ### Changed